Refactor common router integration
